Fixed arm compiler warnings

+++ src/graphics/Framebuffer.cpp
@@ -69,15 +69,15 @@ Framebuffer::Framebuffer(const char *apDevPath)
     // calculate size of screen
     unsigned long screensize = static_cast<unsigned long>(mVariableInfo.yres) * mFixedInfo.line_length;
 
-    mpFrontBuffer = static_cast<uint8_t *>(mmap(0, static_cast<size_t>(screensize * 2), PROT_READ | PROT_WRITE, MAP_SHARED, mFramebufferFile, static_cast<off_t>(0)));
-    if (mpFrontBuffer == reinterpret_cast<uint8_t *>(-1)) /*MAP_FAILED*/ {
+    mpFrontBuffer = static_cast<uint32_t *>(mmap(0, static_cast<size_t>(screensize * 2), PROT_READ | PROT_WRITE, MAP_SHARED, mFramebufferFile, static_cast<off_t>(0)));
+    if (mpFrontBuffer == reinterpret_cast<uint32_t *>(-1)) /*MAP_FAILED*/ {
         THROW_SYSTEM("Framebuffer shared memory mapping failed");
     }
 
     mpBackBuffer = mpFrontBuffer + screensize;
 
     if (mVariableInfo.yoffset > 0) {
-        uint8_t *tmp = mpFrontBuffer;
+        uint32_t *tmp = mpFrontBuffer;
         mpFrontBuffer = mpBackBuffer;
         mpBackBuffer = tmp;
     }
@@ -112,7 +112,7 @@ void Framebuffer::SwapBuffer(const SwapOperations aSwapOp, Color aColor)
     }
 
     // update pointers
-    uint8_t *tmp = mpFrontBuffer;
+    uint32_t *tmp = mpFrontBuffer;
     mpFrontBuffer = mpBackBuffer;
     mpBackBuffer = tmp;
 
@@ -146,12 +146,12 @@ uint32_t Framebuffer::GetPixel(const Point &aPoint, const bool aFront) const
 
 void Framebuffer::clear(Color aColor)
 {
-    long x, y;
+    unsigned long x, y;
     // draw to back buffer
     //     std::cout << "Clearing buffer" << std::endl;
     for (y = 0; y < mVariableInfo.yres; y++) {
         for (x = 0; x < mVariableInfo.xres; x++) {
-            long location = (x + mVariableInfo.xoffset) * (mVariableInfo.bits_per_pixel / 8) + y * mFixedInfo.line_length;
+            unsigned long location = (x + mVariableInfo.xoffset) * (mVariableInfo.bits_per_pixel / 8) + y * mFixedInfo.line_length;
             *(reinterpret_cast<uint32_t *>(mpBackBuffer + location)) = aColor;
         }
     }
@@ -159,11 +159,11 @@ void Framebuffer::clear(Color aColor)
 
 void Framebuffer::copy()
 {
-    long x, y;
+    unsigned long x, y;
     // copy front buffer to back buffer
     for (y = 0; y < mVariableInfo.yres; y++) {
         for (x = 0; x < mVariableInfo.xres; x++) {
-            long location = (x + mVariableInfo.xoffset) * (mVariableInfo.bits_per_pixel / 8) + y * mFixedInfo.line_length;
+            unsigned long location = (x + mVariableInfo.xoffset) * (mVariableInfo.bits_per_pixel / 8) + y * mFixedInfo.line_length;
             *(reinterpret_cast<uint32_t *>(mpBackBuffer + location)) = *(reinterpret_cast<uint32_t *>(mpFrontBuffer + location));
         }
     }
